vision-agent/requirements.txt
railtracks
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
pydantic
